fix(KEN-899): the global scope is used only where it is the person's own
The global branch of the hook lookup never asks about Pi's project trust, and
that was deliberate: it holds the person's own scripts, not a checkout's. Round
4 made the carrier resolve an empty or relative PI_CODING_AGENT_DIR against the
process cwd, to match the renderer. Together those meant Pi launched inside a
checkout nobody had trusted would find that checkout's own
kendex/hooks/<name>.sh through the global branch and spawn it. Spawning a hook
is executing it, so the trust gate added earlier in this branch was reachable
around rather than through. Copilot found it.

The exemption is now conditional on the thing it always assumed. The global
scope is acted on only where the variable names a directory of its own: unset,
or absolute. Where the workspace is untrusted the root must also fall outside
that workspace, since pointing the variable into a clone is not trusting the
clone. Two path comparisons and nothing else, no uid read and no parent walk:
ownership past those questions would be a second security model in the carrier.

Resolution is untouched, so round 4 stands. piUserDir still follows the
renderer wherever the variable points, cwd included, because that is the right
answer for naming a path. Whether to read or run what is there is a different
question, and it is the one that was missing. Both halves are pinned: the
fixtures assert the script really is sitting under the resolved root, then
assert it is not spawned.

The same door opens by reading rather than running, so it is shut in the same
place. The user scope's settings.json is merged unconditionally, so a checkout
sitting where an empty value roots that scope could ship one switching every
guard off, with no code execution needed to reach the same end. Both callers
go through one test.

Withholding is the safe direction on both. An absent hook reads as one kendex
did not install here, which is the answer this carrier already gives. Absent
settings leave the defaults, and every default in this package is on.

No changelog. The carrier spawning rendered scripts, the trust gate and the
PI_CODING_AGENT_DIR reading all arrived in this branch and none has shipped, so
there is no released behaviour to correct and no released hole to report. The
unreleased pi-hooks entry, its README and the Pi adapter doc all said the global
root answers either way, and all three are qualified here rather than left
defending an exemption the code no longer grants unconditionally.

The four fixtures pushed tests/hooks.test.ts to 875 lines, past the 800-line
ratchet for its class. Tests are a frozen class and the house rule is to split
rather than raise, and the seam was already there: which hook the carrier is
allowed to find is a different question from what it does with one it found.
That block is now tests/hook-resolution.test.ts, with the fixture repositories
both suites build in tests/harness.ts. No baseline row, and the same 47 tests
pass across the four files.
