+++ pi-extensions/pi-hooks/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### 0.8.0
 
-- **Breaking**: the three bash guards are no longer reimplemented in TypeScript. `blockBareCd`, `blockRepoCopy` and `preCommitCheck` now spawn the rendered `<project>/.pi/kendex/hooks/<name>.sh`, falling back to `<PI_CODING_AGENT_DIR or ~/.pi/agent>/kendex/hooks/<name>.sh`, with the payload Claude Code sends a `PreToolUse` hook, and read the exit status: `2` refuses the tool call with the script's stderr as the reason, `0` allows it, and any other status refuses too, since a guard that did not run does not stand aside. A script writing to stderr and still exiting `0` is an advisory, shown through the UI and never to the agent. A project-scope script runs only in a workspace Pi reports trusted, since spawning it executes code the project ships; the global root answers either way. The project is the nearest ancestor holding a `.pi/`, `.git/` or `.kendex-lock.json`, so a session started in a subdirectory gets the same hooks as one at the repository root. A name neither root holds runs nothing. The npm package ships the carrier, not the scripts, so install the hooks with `kendex add --hook block-bare-cd --hook block-repo-copy --hook pre-commit-check`. Claude Code, Codex and Pi now run the same bytes, so these guards can no longer disagree.
+- **Breaking**: the three bash guards are no longer reimplemented in TypeScript. `blockBareCd`, `blockRepoCopy` and `preCommitCheck` now spawn the rendered `<project>/.pi/kendex/hooks/<name>.sh`, falling back to `<PI_CODING_AGENT_DIR or ~/.pi/agent>/kendex/hooks/<name>.sh`, with the payload Claude Code sends a `PreToolUse` hook, and read the exit status: `2` refuses the tool call with the script's stderr as the reason, `0` allows it, and any other status refuses too, since a guard that did not run does not stand aside. A script writing to stderr and still exiting `0` is an advisory, shown through the UI and never to the agent. A project-scope script runs only in a workspace Pi reports trusted, since spawning it executes code the project ships; the global root answers either way, but only when `PI_CODING_AGENT_DIR` is unset or absolute, and in an untrusted workspace only when it falls outside that workspace. An empty or relative value roots the global scope at the session's own directory, where a checkout's script would be reached through the branch that skips the trust gate. The same test decides whether this package's user-scope settings are read from that root. The project is the nearest ancestor holding a `.pi/`, `.git/` or `.kendex-lock.json`, so a session started in a subdirectory gets the same hooks as one at the repository root. A name neither root holds runs nothing. The npm package ships the carrier, not the scripts, so install the hooks with `kendex add --hook block-bare-cd --hook block-repo-copy --hook pre-commit-check`. Claude Code, Codex and Pi now run the same bytes, so these guards can no longer disagree.
 - **Breaking**: `extensions/bash-guards.ts` and `extensions/repo-copy-guard.ts` are removed, with `scanCommand`, `preCommitGate`, `CommandScan`, `PreCommitVerdict`, `isBareCd`, `repoCopyRefusal`, `refusalReason`, `isScratch`, `dangerousMarkers` and `classifySegment`. `extensions/process.ts` `runCommandAsync` takes an optional fifth argument, a string written to the child's stdin.
 - **Breaking**: the pre-commit gate reads whitespace-separated words rather than shell, following `hooks/pre-commit-check.sh`. A `git` word with a later `commit` word is the commit, and the bypass is the no-verify flag, a short cluster holding its letter, or a word carrying a `core.hooksPath` key (an attached -c value, the value after a bare -c, a --config-env, a `git config` argument, a `GIT_CONFIG_*` assignment). Nothing else about `-c` is read, so `git commit -c HEAD` and a `user.name` key now pass. Quote characters are dropped and redirection operators become spaces before the words are read, so `git commit "--no-verify"`, `g''it commit --no-verify` and `git commit>/dev/null -n` refuse as the bare flag does. One of those words written inside a commit message, a heredoc body or a comment tail now refuses the command, and the refusal names the rewrite; so does a quoted `commit` in prose, so `git log | grep 'commit'` refuses in a repository nothing armed. A bypass assembled any other way, through an alias, a variable or a key reached through an `include.path`, is not seen. Git's armed hooks are the control either way.
